% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{col}
\alias{col}
\title{colors for plotting}
\format{
A unnamed character string
}
\source{
K. Holsman et al. 2024
}
\usage{
col
}
\description{
Three colors for color ramp plotting
}
\keyword{datasets}
