% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_model.R
\name{compile_model}
\alias{compile_model}
\title{compile_model}
\usage{
compile_model(src_dir = "src/TMB", recompile = T, model = "VonBEEv1")
}
\arguments{
\item{src_dir}{source of the model "src/TMB"}

\item{recompile}{T/F recompile the model}

\item{model}{model version to compile, default is "VonBEEv1"}
}
\description{
compile or recompile from the original .cpp file
}
