% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diet_p}
\alias{diet_p}
\title{Diet data and plots for the EBS}
\format{
A list with three objects
\describe{
  \item{data}{diet proportions by weight of key prey items in the diet of each predator}
  \item{plot}{Diet proportion plot}
  \item{prey}{Colors for each prey item for plotting}
}
}
\source{
K. Holsman et al. 2024
}
\usage{
diet_p
}
\description{
A data set with the scores of two groups.
}
\keyword{datasets}
