% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_fun_makeMap.R
\name{makeMap}
\alias{makeMap}
\title{makeMap will create a map list for reading into the runmod() function}
\usage{
makeMap(param = parameters, estpar = estparams)
}
\arguments{
\item{estpar}{vector of T/F of which parameters will be estimated}

\item{parameters}{list of starting values and parameters for the TMB futR() model}
}
\value{
maplist a list with map values
}
\description{
This function prepares the estimated data for futR
This function is a subset of makeDat()
}
\examples{

datlist   <- readMake_futR_data("data/in/futR_Inputs.xlsx" )

estparams <- datlist$estparams
estparams

parameters <- atlist$parameters
parameters
maplist <- makeMap(param=parameters,estpar=estparams)

}
