% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mnLWA}
\alias{mnLWA}
\title{mean Length Weight and Age data for Bering Sea walleye pollock}
\format{
A dataframe with the following columns
\describe{
\item{REGION}{Alaska basing (GOA, AI, or BS)}
\item{YEAR}{Year of the survey}
\item{STRATUM}{Survey stratum}
\item{LAT}{Survey station LAT}
\item{LON}{Survey station LAT}
\item{CRUISE_TYPE}{Specifying the survey cruise}
\item{SPECIES_NAME}{Internal code to match species info}
\item{COMMON_NAME}{Common name}
\item{SPECIES_NAME}{Species name}
\item{mnWEIGHT}{mean Weight of the fish in grams}
\item{mnLENGTH}{mean Length of the fish in mm}
\item{mnAGE}{mean Age of the fish in years}
\item{mnSST}{mean sea surface temperature at the survey location}
\item{mnBT}{mean bottom temperature at the survey location}
\item{sdWEIGHT}{standard deviation of the Weight of the fish in grams}
\item{sdLENGTH}{standard deviation of the  Length of the fish in mm}
\item{sdAGE}{standard deviation of the  Age of the fish in years}
\item{sdSST}{standard deviation of the  sea surface temperature at the survey location}
\item{sdBT}{standard deviation of the  bottom temperature at the survey location}
\item{nobs}{number of observations for the sample}

}
}
\source{
K. Holsman et al. 2024
}
\usage{
mnLWA
}
\description{
Length Weight and Age data from summer ground surveys conducted by the NOAA NMFS Alaska Fisheries Science Center
}
\keyword{datasets}
