% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_fun_runMod.R
\name{runMod}
\alias{runMod}
\title{run the vonBT()  model}
\usage{
runMod(
  dlistIN,
  src_fldr = "src",
  version = "vonBT",
  nrep = 3,
  optimizer = "nlminb",
  convergenceTest = TRUE,
  methodIN = NULL,
  hessianIN = TRUE,
  silentIN = TRUE,
  se.fit = TRUE,
  simulate = FALSE,
  sim_nitr = 1000,
  recompile = FALSE,
  maxitr = 10000,
  maxeval = 10000
)
}
\arguments{
\item{dlistIN}{list of input data and parameters used for the makeADfun() dependency
* dlistIN$parameters  list of parameters for the TMB vonBT model:
* dlistIN$randonm   list of which parms should be radnom effects
* dlistIN$data      list of data to read into the model
* dlistIN$map       list of the mapp for the TMB model
* dlistIN$phases    (defunct) phases for estimating parameters
* dlistIN$inputs    archive of intputs for this function}

\item{src_fldr}{subfolder where model is located. Default is 'src'}

\item{version}{recruitment model to run. Default is 'vonBT'}

\item{optimizer}{default is "nlminb"; can also be set to "optim"}

\item{convergenceTest}{Test for failed convergence by also running the model using optim}

\item{methodIN}{default is NULL, can be set to "Nelder-Mead", "BFGS", "CG", "L-BFGS-B", "SANN", or "Brent"}

\item{silentIN}{T/F , default = 'TRUE'}

\item{se.fit}{T/F , default = 'TRUE'}

\item{simulate}{T/F , simulate the model? default = 'FALSE'}

\item{sim_nitr}{number of iterations for the simulate function}

\item{recompile}{recompile the .cpp file? default is 'FALSE'}

\item{maxitr}{10000  Max iterations for fitting the objective function}

\item{maxeval}{10000  Max evaluations for fitting the objective function}
}
\value{
returns  summary of the model including the mle
}
\description{
runMod() will run the vonBT() recruitment model
}
\examples{
datlist <- readMake_vonBT_data("data/in/vonBT_Inputs.xlsx" )
mm      <- runRecMod(dlistIN   = datlist, version   = 'vonBT',recompile = FALSE,simulate  = TRUE,sim_nitr  = 1000)
}
