% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_fun_sim_dat.R
\name{sim_dat}
\alias{sim_dat}
\title{sim_dat()}
\usage{
sim_dat(par, data, simulateIT = FALSE)
}
\arguments{
\item{par}{parameters, e.g., dlistIN_0$par}

\item{data}{data file to simulate from, e.g., datIN}

\item{simulateIT}{TRUE/FALSE}
}
\value{
list(out = cbind(data.frame(W_obs = exp(logWobs),W_hat = What, age=age),t(vonb_cov)),What=What,Wobs=exp(logWobs),age=age, vonb_cov=vonb_cov,d=d, Winf=Winf,nll=nll)
}
\description{
simulate the TMB model in for VonBEE
}
