% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_fun_test_fun.R
\name{test_fun}
\alias{test_fun}
\title{test_fun}
\usage{
test_fun(x, age = 5, t0 = -0.25, K = 12, H = 0.25, offset = 0)
}
\arguments{
\item{age}{=5}

\item{t0}{=-.25}

\item{K}{= 12}

\item{H}{= .25}

\item{offset}{=0}
}
\value{
list(d=d, Winf = Winf,logWhat = logWhat, What=What, parta = parta , partb=partb)
}
\description{
maybe defunct?
}
