% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_fun_sim_dat.R
\name{sim_dat}
\alias{sim_dat}
\title{sim_dat()}
\usage{
sim_dat(par = dlistIN_0$par, data = datIN, simulateIT = FALSE)
}
\arguments{
\item{par}{=dlistIN_0$par}

\item{data}{= datIN}

\item{simulateIT}{=FALSE}
}
\value{
list(out = cbind(data.frame(W_obs = exp(logWobs),W_hat = What, age=age),t(vonb_cov)),What=What,Wobs=exp(logWobs),age=age, vonb_cov=vonb_cov,d=d, Winf=Winf,nll=nll)
}
\description{
simulate the TMB model in for VonBT in R
}
